app/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.29.0
pydantic==2.6.4
python-dotenv==1.0.1
redis==5.0.3
httpx==0.28.1
starlette==0.46.2
celery==5.3.6

# testes da API
pytest==8.3.5
pytest-asyncio==0.26.0